Incorrect return type on a compat header.
This was exposed by the update to musl 0.9.13 which now has a
prototype for this function.

+++ system/include/compat/xlocale.h
@@ -9,7 +9,7 @@ extern "C" {
 
 long long strtoll_l(const char *start, char **end, int base, locale_t loc);
 unsigned long long strtoull_l(const char *start, char **end, int base, locale_t loc);
-double strtold_l(const char *start, char **end, locale_t loc);
+long double strtold_l(const char *start, char **end, locale_t loc);
 
 #ifdef __cplusplus
 }
